requirements.txt
eth-account==0.13.7
Flask==3.1.3
py_clob_client==0.34.6
py_order_utils==0.3.2
python-dotenv==1.2.1
requests==2.32.5
web3==7.14.1
